test(p2-shim): separate browser and node suites
