+++ packages/preview2-shim/test/browser.ts
@@ -1,3 +1,4 @@
+import { rejects, throws } from "node:assert";
 import { writeFile, mkdir, readFile } from "node:fs/promises";
 import { createSecureServer } from "node:http2";
 import { dirname } from "node:path";
@@ -9,6 +10,7 @@ import { transpile } from "@bytecodealliance/jco";
 import { getTmpDir, FIXTURES_WIT_DIR, startTestServer, runBasicHarnessPageTest } from "./common.js";
 
 type TranspileOutput = { files: { [filename: string]: Uint8Array } };
+const symbolDispose = Symbol.dispose || Symbol.for("dispose");
 const LOCALHOST_KEY = new URL("./fixtures/tls/localhost.key", import.meta.url);
 const LOCALHOST_CERTIFICATE = new URL("./fixtures/tls/localhost.crt", import.meta.url);
 
@@ -1195,5 +1197,611 @@ export const test = {
         assert.strictEqual(statusJSON.msg, successMsg);
 
         await cleanup();
+    });
+});
+
+suite("Browser filesystem", () => {
+    test("writeViaStream reuses file capacity", async () => {
+        const { _setFileData, preopens } = await import("../src/browser/filesystem.js");
+        const file = { source: new Uint8Array([1, 2, 3]) };
+        _setFileData({ dir: { file } });
+
+        const [[rootDescriptor]] = preopens.getDirectories();
+        const descriptor = rootDescriptor.openAt({}, "file", {}, { write: true });
+        const stream = descriptor.writeViaStream(3n);
+
+        stream.checkWrite();
+        stream.write(new Uint8Array([4, 5]));
+        const buffer = file.source.buffer;
+        stream.checkWrite();
+        stream.write(new Uint8Array([6]));
+
+        assert.strictEqual(file.source.buffer, buffer);
+        assert.deepStrictEqual(file.source, new Uint8Array([1, 2, 3, 4, 5, 6]));
+    });
+
+    test("writeViaStream preserves overwrite and sparse-write semantics", async () => {
+        const { _setFileData, preopens } = await import("../src/browser/filesystem.js");
+        const file = { source: new Uint8Array([1, 2, 3, 4]) };
+        _setFileData({ dir: { file } });
+
+        const [[rootDescriptor]] = preopens.getDirectories();
+        const descriptor = rootDescriptor.openAt({}, "file", {}, { write: true });
+
+        const overwrite = descriptor.writeViaStream(1n);
+        overwrite.checkWrite();
+        overwrite.write(new Uint8Array([8, 9]));
+        assert.deepStrictEqual(file.source, new Uint8Array([1, 8, 9, 4]));
+
+        const sparse = descriptor.writeViaStream(6n);
+        sparse.checkWrite();
+        sparse.write(new Uint8Array([7]));
+        assert.deepStrictEqual(file.source, new Uint8Array([1, 8, 9, 4, 0, 0, 7]));
+    });
+
+    test("supports append, resize, rename, and removal", async () => {
+        const { _setFileData, preopens } = await import("../src/browser/filesystem.js");
+        const file = { source: new Uint8Array([1, 2]) };
+        _setFileData({ dir: { file, empty: { dir: {} } } });
+
+        const [[root]] = preopens.getDirectories();
+        const descriptor = root.openAt({}, "file", {}, { read: true, write: true });
+        const append = descriptor.appendViaStream();
+        append.checkWrite();
+        append.write(new Uint8Array([3]));
+        assert.deepStrictEqual(file.source, new Uint8Array([1, 2, 3]));
+
+        descriptor.setSize(5n);
+        assert.deepStrictEqual(file.source, new Uint8Array([1, 2, 3, 0, 0]));
+        root.renameAt("file", root, "renamed");
+        assert.strictEqual(root.statAt({}, "renamed").size, 5n);
+        root.unlinkFileAt("renamed");
+        assert.throws(() => root.statAt({}, "renamed"));
+        root.removeDirectoryAt("empty");
+        assert.throws(() => root.statAt({}, "empty"));
+    });
+
+    test("creates only the final path component", async () => {
+        const { _setFileData, preopens } = await import("../src/browser/filesystem.js");
+        const fileData = { dir: { parent: { dir: {} }, existing: { dir: {} } } };
+        _setFileData(fileData);
+        const [[root]] = preopens.getDirectories();
+
+        let error: unknown;
+        try {
+            root.openAt({}, "missing/file", { create: true }, { write: true });
+        } catch (caught) {
+            error = caught;
+        }
+        assert.strictEqual(error, "no-entry");
+        assert.strictEqual((fileData.dir as Record<string, unknown>).missing, undefined);
+
+        const created = root.openAt({}, "parent/file", { create: true }, { write: true });
+        assert.strictEqual(created.getType(), "regular-file");
+        root.createDirectoryAt("parent/child");
+        assert.strictEqual(root.statAt({}, "parent/child").type, "directory");
+        error = undefined;
+        try {
+            root.createDirectoryAt("existing");
+        } catch (caught) {
+            error = caught;
+        }
+        assert.strictEqual(error, "exist");
+    });
+
+    test("renames entries without deleting or corrupting the tree", async () => {
+        const { _setFileData, preopens } = await import("../src/browser/filesystem.js");
+        _setFileData({
+            dir: {
+                file: { source: "source" },
+                target: { source: "target" },
+                directory: { dir: { child: { source: "child" } } },
+                empty: { dir: {} },
+                nonempty: { dir: { value: { source: "value" } } },
+            },
+        });
+        const [[root]] = preopens.getDirectories();
+        const thrownValue = (fn: () => void) => {
+            try {
+                fn();
+            } catch (error) {
+                return error;
+            }
+            assert.fail("operation should have thrown");
+        };
+
+        root.renameAt("file", root, "file");
+        assert.strictEqual(root.statAt({}, "file").type, "regular-file");
+
+        root.renameAt("file", root, "target");
+        assert.strictEqual(root.statAt({}, "target").size, 6n);
+        assert.strictEqual(
+            thrownValue(() => root.statAt({}, "file")),
+            "no-entry",
+        );
+
+        assert.strictEqual(
+            thrownValue(() => root.renameAt("target", root, "empty")),
+            "is-directory",
+        );
+        assert.strictEqual(
+            thrownValue(() => root.renameAt("directory", root, "target")),
+            "not-directory",
+        );
+        assert.strictEqual(
+            thrownValue(() => root.renameAt("directory", root, "nonempty")),
+            "not-empty",
+        );
+
+        const directory = root.openAt({}, "directory", { directory: true }, {});
+        assert.strictEqual(
+            thrownValue(() => root.renameAt("directory", directory, "descendant")),
+            "invalid",
+        );
+        assert.strictEqual(root.statAt({}, "directory/child").type, "regular-file");
+
+        root.renameAt("directory", root, "empty");
+        assert.strictEqual(root.statAt({}, "empty/child").type, "regular-file");
+    });
+
+    test("shares identity and metadata across descriptors and hard links", async () => {
+        const { _setFileData, preopens } = await import("../src/browser/filesystem.js");
+        _setFileData({ dir: { file: { source: "value" } } });
+        const [[root]] = preopens.getDirectories();
+        const first = root.openAt({}, "file", {}, { read: true, write: true });
+        const second = root.openAt({}, "file", {}, { read: true, write: true });
+
+        assert.strictEqual(first.isSameObject(second), true);
+        assert.deepStrictEqual(root.metadataHashAt({}, "file"), first.metadataHash());
+        const before = first.metadataHash();
+        second.write(new Uint8Array([1]), 0n);
+        assert.notDeepEqual(first.metadataHash(), before);
+        assert.deepStrictEqual(first.metadataHash(), second.metadataHash());
+
+        root.linkAt({}, "file", root, "link");
+        const link = root.openAt({}, "link", {}, { read: true });
+        assert.strictEqual(first.isSameObject(link), true);
+        assert.strictEqual(first.stat().linkCount, 2n);
+        assert.strictEqual(link.stat().linkCount, 2n);
+        root.unlinkFileAt("file");
+        assert.strictEqual(link.stat().linkCount, 1n);
+        assert.deepStrictEqual(root.metadataHashAt({}, "link"), link.metadataHash());
+    });
+
+    test("createFilesystem isolates explicitly selected in-memory roots", async () => {
+        const { createFilesystem, InMemoryFilesystemAdapter } =
+            await import("../src/browser/filesystem.js");
+        const first = createFilesystem({
+            adapter: new InMemoryFilesystemAdapter(),
+            preopens: { "/first": { dir: { value: { source: "one" } } } },
+        });
+        const second = createFilesystem({
+            adapter: new InMemoryFilesystemAdapter(),
+            preopens: { "/second": { dir: { value: { source: "two" } } } },
+        });
+
+        assert.strictEqual(first.preopens.getDirectories()[0][1], "/first");
+        assert.strictEqual(second.preopens.getDirectories()[0][1], "/second");
+        first.dispose();
+        assert.throws(() => first.preopens.getDirectories(), /disposed/);
+        assert.strictEqual(second.preopens.getDirectories().length, 1);
+    });
+});
+
+suite("Browser shim guards", () => {
+    test("pollList throws on empty list", async () => {
+        const { poll } = await import("../src/browser/io.js");
+        assert.throws(() => poll.poll([]), /empty/);
+    });
+
+    test("pollList throws on list exceeding u32 range", async () => {
+        const { poll } = await import("../src/browser/io.js");
+        const fakeList = { length: 0x100000000 } as any;
+        assert.throws(() => poll.poll(fakeList), /u32/);
+    });
+
+    test("pollables can be reused across readiness generations", async () => {
+        const { pollableCreate } = await import("../src/browser/io.js");
+        let ready = false;
+        let waits = 0;
+        let resolve!: () => void;
+        const pollable = pollableCreate({
+            ready: () => ready,
+            wait: () => {
+                waits++;
+                return new Promise<void>((r) => (resolve = r));
+            },
+        });
+
+        const first = pollable.block();
+        const simultaneous = pollable.block();
+        assert.strictEqual(waits, 1);
+        ready = true;
+        resolve();
+        await Promise.all([first, simultaneous]);
+        assert.strictEqual(pollable.ready(), true);
+
+        ready = false;
+        const second = pollable.block();
+        assert.strictEqual(waits, 2);
+        ready = true;
+        resolve();
+        await second;
+    });
+
+    test("poll reuses pollables and returns ready indices in input order", async () => {
+        const { poll, pollableCreate } = await import("../src/browser/io.js");
+        let firstReady = false;
+        let secondReady = true;
+        let resolveFirst!: () => void;
+        let resolveSecond!: () => void;
+        const first = pollableCreate({
+            ready: () => firstReady,
+            wait: () => new Promise<void>((resolve) => (resolveFirst = resolve)),
+        });
+        const second = pollableCreate({
+            ready: () => secondReady,
+            wait: () => new Promise<void>((resolve) => (resolveSecond = resolve)),
+        });
+
+        assert.deepStrictEqual(await poll.poll([first, second, first]), new Uint32Array([1]));
+        secondReady = false;
+        const next = poll.poll([first, second, first]);
+        firstReady = true;
+        resolveFirst();
+        assert.deepStrictEqual(await next, new Uint32Array([0, 2]));
+
+        firstReady = false;
+        const final = poll.poll([first, second]);
+        secondReady = true;
+        resolveSecond();
+        assert.deepStrictEqual(await final, new Uint32Array([1]));
+    });
+
+    test("browser streams validate u64 lengths and write permits", async () => {
+        const { inputStreamCreate, outputStreamCreate } = await import("../src/browser/io.js");
+        const input = inputStreamCreate({ blockingRead: () => new Uint8Array() });
+        assert.throws(() => input.read(-1n), /valid u64/);
+        assert.throws(() => input.read(BigInt(Number.MAX_SAFE_INTEGER) + 1n), /safe integer/);
+
+        const output = outputStreamCreate({
+            checkWrite: () => 2n,
+            write() {},
+        });
+        assert.strictEqual(output.checkWrite(), 2n);
+        assert.throws(() => output.write(new Uint8Array(3)), /exceeds the permit/);
+        assert.throws(() => output.writeZeroes(3n), /exceeds the permit/);
+        assert.throws(() => output.blockingWriteZeroesAndFlush(4097n), /at most 4096/);
+    });
+
+    test("browser blocking stream fallbacks flush and accept external pollables", async () => {
+        const { inputStreamCreate, outputStreamCreate } = await import("../src/browser/io.js");
+        const events: string[] = [];
+        const output = outputStreamCreate({
+            write: () => events.push("write"),
+            flush: () => events.push("flush"),
+        });
+        output.blockingWriteAndFlush(new Uint8Array([1]));
+        output.blockingFlush();
+        assert.deepStrictEqual(events, ["write", "flush", "flush"]);
+
+        const externalPollable = { ready: () => true, block() {} };
+        const input = inputStreamCreate({
+            blockingRead: () => new Uint8Array(),
+            subscribe: () => externalPollable as any,
+        });
+        assert.strictEqual(input.subscribe(), externalPollable);
+        (input as any)[symbolDispose]();
+    });
+
+    test("browser clocks validate u64 subscriptions and make elapsed timers ready", async () => {
+        const { monotonicClock, wallClock } = await import("../src/browser/clocks.js");
+        assert.strictEqual(typeof monotonicClock.resolution(), "bigint");
+        assert.strictEqual(typeof wallClock.resolution().seconds, "bigint");
+        assert.strictEqual(typeof wallClock.resolution().nanoseconds, "number");
+        assert.strictEqual(monotonicClock.subscribeDuration(0n).ready(), true);
+        assert.strictEqual(monotonicClock.subscribeInstant(monotonicClock.now()).ready(), true);
+        assert.throws(() => monotonicClock.subscribeDuration(-1n), /valid u64/);
+        assert.throws(() => monotonicClock.subscribeInstant(-1n), /valid u64/);
+    });
+
+    test("dropping browser streams disposes handlers exactly once", async () => {
+        const { outputStreamCreate } = await import("../src/browser/io.js");
+        let drops = 0;
+        const output = outputStreamCreate({ write() {}, drop: () => drops++ });
+        (output as any)[symbolDispose]();
+        (output as any)[symbolDispose]();
+        assert.strictEqual(drops, 1);
+        try {
+            output.checkWrite();
+            assert.fail("closed output stream should reject checkWrite");
+        } catch (error) {
+            assert.deepStrictEqual(error, { tag: "closed" });
+        }
+    });
+
+    test("dropping resources wakes blocked child pollables", async () => {
+        const { inputStreamCreate, poll, pollableCreate } = await import("../src/browser/io.js");
+        const input = inputStreamCreate({
+            blockingRead: () => new Uint8Array(),
+            subscribe: () =>
+                // The source intentionally never becomes ready on its own.
+                pollableCreate({
+                    ready: () => false,
+                    wait: () => new Promise<void>(() => {}),
+                }),
+        });
+        const child = input.subscribe();
+        const blocked = child.block();
+        const polled = poll.poll([child]) as Promise<Uint32Array>;
+
+        (input as any)[symbolDispose]();
+
+        const blockError = await blocked.then(
+            () => undefined,
+            (error) => error,
+        );
+        const pollError = await polled.then(
+            () => undefined,
+            (error) => error,
+        );
+        assert.match(blockError.message, /parent resource has been disposed/);
+        assert.match(pollError.message, /parent resource has been disposed/);
+    });
+
+    test("browser random rejects invalid allocation lengths", async () => {
+        const { random } = await import("../src/browser/random.js");
+        assert.throws(() => random.getRandomBytes(-1n), /valid u64/);
+        assert.throws(
+            () => random.getRandomBytes(BigInt(Number.MAX_SAFE_INTEGER) + 1n),
+            /safe integer/,
+        );
+    });
+
+    test("browser raw sockets fail with not-supported", async () => {
+        const sockets = await import("../src/browser/sockets.js");
+        assert.strictEqual(
+            sockets.instanceNetwork.instanceNetwork(),
+            sockets.instanceNetwork.instanceNetwork(),
+        );
+        assert.throws(() => sockets.tcpCreateSocket.createTcpSocket("ipv4"), /not-supported/);
+        assert.throws(() => sockets.udpCreateSocket.createUdpSocket("ipv4"), /not-supported/);
+        assert.throws(
+            () => sockets.ipNameLookup.resolveAddresses({} as any, "example.com"),
+            /not-supported/,
+        );
+    });
+
+    test("host-driven browser incoming HTTP round trips a response", async () => {
+        const { handleIncomingRequest, types } = await import("../src/browser/http.js");
+        const response = await handleIncomingRequest(
+            new Request("https://example.com/test?value=1", { method: "POST", body: "request" }),
+            (request, responseOut) => {
+                assert.deepStrictEqual(request.method(), { tag: "post" });
+                assert.strictEqual(request.pathWithQuery(), "/test?value=1");
+                const outgoing = new types.OutgoingResponse(new types.Fields());
+                outgoing.setStatusCode(201);
+                const body = outgoing.body();
+                const stream = body.write();
+                stream.checkWrite();
+                stream.write(new TextEncoder().encode("created"));
+                types.OutgoingBody.finish(body, undefined);
+                types.ResponseOutparam.set(responseOut, { tag: "ok", val: outgoing });
+            },
+        );
+        assert.strictEqual(response.status, 201);
+        assert.strictEqual(await response.text(), "created");
+    });
+
+    test("browser incoming HTTP rejects an unset response outparam", async () => {
+        const { handleIncomingRequest } = await import("../src/browser/http.js");
+        await rejects(
+            handleIncomingRequest(new Request("https://example.com/"), () => {}),
+            /without setting its response outparam/,
+        );
+    });
+
+    test("browser incoming HTTP exposes an injectable handler namespace", async () => {
+        const { createIncomingHandler, types } = await import("../src/browser/http.js");
+        let called = false;
+        const handler = createIncomingHandler((_request, responseOut) => {
+            called = true;
+            types.ResponseOutparam.set(responseOut, {
+                tag: "err",
+                val: { tag: "internal-error", val: "rejected" },
+            });
+        });
+        const response = await handleIncomingViaNamespace(handler);
+        assert.strictEqual(called, true);
+        assert.strictEqual(response.status, 500);
+        assert.match(await response.text(), /internal-error.*rejected/);
+
+        async function handleIncomingViaNamespace(namespace: typeof handler) {
+            const { handleIncomingRequest } = await import("../src/browser/http.js");
+            return handleIncomingRequest(new Request("https://example.com/"), namespace.handle);
+        }
+    });
+
+    test("browser outgoing HTTP waits for the complete request body", async () => {
+        const { _setRequestStreaming, outgoingHandler, types } =
+            await import("../src/browser/http.js");
+        _setRequestStreaming(false);
+        const originalFetch = globalThis.fetch;
+        let requestedBody: Uint8Array | undefined;
+        let fetchCalls = 0;
+        globalThis.fetch = async (_input, init) => {
+            fetchCalls++;
+            requestedBody = new Uint8Array(await new Response(init?.body).arrayBuffer());
+            return new Response("ok");
+        };
+        try {
+            const request = new types.OutgoingRequest(new types.Fields());
+            request.setMethod({ tag: "post" });
+            request.setScheme({ tag: "HTTPS" });
+            request.setAuthority("example.com");
+            request.setPathWithQuery("/");
+            const body = request.body();
+            const stream = body.write();
+
+            const response = outgoingHandler.handle(request, undefined);
+            await Promise.resolve();
+            assert.strictEqual(fetchCalls, 0);
+
+            stream.checkWrite();
+            stream.write(new TextEncoder().encode("complete body"));
+            types.OutgoingBody.finish(body, undefined);
+            await response.subscribe().block();
+
+            assert.strictEqual(fetchCalls, 1);
+            assert.strictEqual(new TextDecoder().decode(requestedBody), "complete body");
+            throws(
+                () => stream.write(new Uint8Array([0])),
+                ({ tag }) => tag === "closed",
+            );
+        } finally {
+            globalThis.fetch = originalFetch;
+        }
+    });
+
+    test("browser outgoing HTTP can stream request bodies to Fetch", async () => {
+        const { _setRequestStreaming, outgoingHandler, types } =
+            await import("../src/browser/http.js");
+        const originalFetch = globalThis.fetch;
+        let fetchCalls = 0;
+        let duplex: string | undefined;
+        let bodyText: Promise<string> | undefined;
+        globalThis.fetch = async (_input, init) => {
+            fetchCalls++;
+            duplex = (init as RequestInit & { duplex?: string }).duplex;
+            bodyText = new Response(init?.body).text();
+            return new Response("ok");
+        };
+        _setRequestStreaming(true);
+        try {
+            const request = new types.OutgoingRequest(new types.Fields());
+            request.setMethod({ tag: "post" });
+            request.setScheme({ tag: "HTTPS" });
+            request.setAuthority("example.com");
+            request.setPathWithQuery("/");
+            const body = request.body();
+            const stream = body.write();
+            stream.checkWrite();
+            stream.write(new TextEncoder().encode("before "));
+
+            const response = outgoingHandler.handle(request, undefined);
+            await Promise.resolve();
+            assert.strictEqual(fetchCalls, 1);
+            assert.strictEqual(duplex, "half");
+
+            stream.checkWrite();
+            stream.write(new TextEncoder().encode("finish"));
+            types.OutgoingBody.finish(body, undefined);
+            assert.strictEqual(await bodyText, "before finish");
+            await response.subscribe().block();
+        } finally {
+            _setRequestStreaming(false);
+            globalThis.fetch = originalFetch;
+        }
+    });
+
+    test("browser CLI factories isolate configuration and streams", async () => {
+        const { createCli } = await import("../src/browser/cli.js");
+        const firstWrites: number[] = [];
+        const secondWrites: number[] = [];
+        const first = createCli({
+            environment: { INSTANCE: "first" },
+            arguments: ["one"],
+            stdout: { write: (bytes) => firstWrites.push(...bytes) },
+        });
+        const second = createCli({
+            environment: { INSTANCE: "second" },
+            arguments: ["two"],
+            stdout: { write: (bytes) => secondWrites.push(...bytes) },
+        });
+
+        assert.deepStrictEqual(first.environment.getEnvironment(), [["INSTANCE", "first"]]);
+        assert.deepStrictEqual(second.environment.getEnvironment(), [["INSTANCE", "second"]]);
+        const firstStdout = first.stdout.getStdout();
+        firstStdout.checkWrite();
+        firstStdout.write(new Uint8Array([1, 2]));
+        assert.deepStrictEqual(firstWrites, [1, 2]);
+        assert.deepStrictEqual(secondWrites, []);
+        assert.strictEqual(first.terminalStdout.getTerminalStdout(), undefined);
+    });
+
+    test("RequestOptions rejects negative connect timeout", async () => {
+        const { types } = await import("../src/browser/http.js");
+        const opts = new types.RequestOptions();
+        assert.throws(() => opts.setConnectTimeout(-1n), /negative/);
+    });
+
+    test("RequestOptions rejects negative first-byte timeout", async () => {
+        const { types } = await import("../src/browser/http.js");
+        const opts = new types.RequestOptions();
+        assert.throws(() => opts.setFirstByteTimeout(-1n), /negative/);
+    });
+
+    test("RequestOptions rejects negative between-bytes timeout", async () => {
+        const { types } = await import("../src/browser/http.js");
+        const opts = new types.RequestOptions();
+        assert.throws(() => opts.setBetweenBytesTimeout(-1n), /negative/);
+    });
+
+    test("browser HTTP options preserve undefined timeouts", async () => {
+        const { types } = await import("../src/browser/http.js");
+        const opts = new types.RequestOptions();
+        assert.strictEqual(opts.connectTimeout(), undefined);
+        opts.setConnectTimeout(10n);
+        opts.setConnectTimeout(undefined);
+        opts.setFirstByteTimeout(undefined);
+        opts.setBetweenBytesTimeout(undefined);
+        assert.strictEqual(opts.connectTimeout(), undefined);
+        assert.strictEqual(opts.firstByteTimeout(), undefined);
+        assert.strictEqual(opts.betweenBytesTimeout(), undefined);
+    });
+
+    test("browser HTTP validates DNS, IPv4, and IPv6 authorities", async () => {
+        const { types } = await import("../src/browser/http.js");
+        const request = new types.OutgoingRequest(new types.Fields());
+        for (const authority of ["example.com", "127.0.0.1:8080", "[::1]:8080"]) {
+            request.setAuthority(authority);
+            assert.strictEqual(request.authority(), authority);
+        }
+        for (const authority of [
+            "",
+            "user@example.com",
+            "example.com:",
+            "example.com:65536",
+            "::1",
+            "[not-ipv6]",
+        ]) {
+            throws(() => request.setAuthority(authority));
+        }
+    });
+
+    test("browser HTTP preserves repeated outgoing headers", async () => {
+        const { outgoingHandler, types } = await import("../src/browser/http.js");
+        const originalFetch = globalThis.fetch;
+        let receivedHeader: string | null = null;
+        globalThis.fetch = async (_input, init) => {
+            receivedHeader = (init?.headers as Headers).get("x-repeat");
+            return new Response();
+        };
+        try {
+            const encoder = new TextEncoder();
+            const request = new types.OutgoingRequest(
+                types.Fields.fromList([
+                    ["x-repeat", encoder.encode("one")],
+                    ["x-repeat", encoder.encode("two")],
+                ]),
+            );
+            request.setAuthority("example.com");
+            request.setPathWithQuery("/");
+            const response = outgoingHandler.handle(request, undefined);
+            await response.subscribe().block();
+            assert.strictEqual(receivedHeader, "one, two");
+        } finally {
+            globalThis.fetch = originalFetch;
+        }
     });
 });
